requirements.txt
pandas>=1.5.0
numpy>=1.23.0
statsmodels>=0.14.0
scikit-learn>=1.2.0
lightgbm>=4.0.0
matplotlib>=3.6.0
pyyaml>=6.0
optuna>=3.0.0
